sync existing course from canvas
soc info still missing
